Fixing json dumps bug and other minor fixes
Converting the initial dictionary back into a string using JSON dumps
resulted in a ValueError exception when attempting to convert that same
string back to a dictionary using ast.literal_eval since booleans in
JSON are lowercase true/false. Doing a straight cast avoids this
problem. Thanks keekz!

Updated the list of acceptable packageType's to reflect what is more
accurately accepted. The API docs are not 100% in sync with what
Artifactory expects.

Fixed a copy/paste typo where repository messages were being reported as
group messages.

Removed 'yum' as an accepted packageType for repositories since
Artifactory automatically converts the repository to 'rpm' anyway. This
was causing a bug where the config was always out of sync.