Add urllib.parse.quote when creating working_url
Currently, working url takes the name provided by the ansible module. If
this name contains spaces, however, it'll result in an error. As it
should. This upgrade takes the "name" provided from ansible modules and
escapes the characters so that we can make calls properly.

For the artifactory_groups and artifactory_permissions  module, we need to
bring the "name" in the final dict to lowercase since artifactory brings all
names of LDAP groups to lowercase and compares against that.